+++ bootstrappers/bootstrap-jdk.sh
@@ -21,38 +21,35 @@ main() {
     read -p "$prefix Enter [y|n]:" answer
     case $answer in
       [Yy])
-        if ! type -t brew
+        if [[ $(type -t brew) = "" ]] && [ ! -f /opt/homebrew/bin/brew ]
         then
-          if [[ ! -e /opt/homebrew/bin/brew ]]
-          then
             info "Cannot run this bootstrapper. Install required dependency first: Homebrew."
-          else
-            eval "$(/opt/homebrew/bin/brew shellenv)"
+        else
+          eval "$(/opt/homebrew/bin/brew shellenv)"
 
-            brew install openjdk@11
-            brew install jenv
+          brew install openjdk@11
+          brew install jenv
 
-            echo '### jEnv' >> $HOME/env.sh
-            echo 'export PATH="$HOME/.jenv/bin:$PATH"' >> $HOME/env.sh
-            echo 'eval "$(jenv init -)"' >> $HOME/env.sh
+          echo '### jEnv' >> $HOME/env.sh
+          echo 'export PATH="$HOME/.jenv/bin:$PATH"' >> $HOME/env.sh
+          echo 'eval "$(jenv init -)"' >> $HOME/env.sh
 
-            source $HOME/env.sh
+          source $HOME/env.sh
 
-            # If the ~/.jenv/versions folder is empty, add all of the current JDK installations into jenv.
-            if [ -z "$(find "${HOME}/.jenv/versions" -mindepth 1 -maxdepth 1 -print -quit 2> /dev/null)" ]; then
-              while IFS= read -r -d $'\n' jdk; do
-                if [ -d "${jdk}/Contents/Home/bin" ]; then
-                  jenv add "${jdk}/Contents/Home";
-                fi;
-              done < <(find /Library/Java/JavaVirtualMachines -type d -maxdepth 1 -mindepth 1 -print);
-            fi;
+          # If the ~/.jenv/versions folder is empty, add all of the current JDK installations into jenv.
+          if [ -z "$(find "${HOME}/.jenv/versions" -mindepth 1 -maxdepth 1 -print -quit 2> /dev/null)" ]; then
+            while IFS= read -r -d $'\n' jdk; do
+              if [ -d "${jdk}/Contents/Home/bin" ]; then
+                jenv add "${jdk}/Contents/Home";
+              fi;
+            done < <(find /Library/Java/JavaVirtualMachines -type d -maxdepth 1 -mindepth 1 -print);
+          fi;
 
-            # Default to Java 11.
-            if [ ! -r "${HOME}/.jenv/version" ]; then
-              jenv global 11;
-              jenv rehash;
-            fi;
-          fi
+          # Default to Java 11.
+          if [ ! -r "${HOME}/.jenv/version" ]; then
+            jenv global 11;
+            jenv rehash;
+          fi;
         fi
         break;
       ;;
