+++ bootstrappers/bootstrap-vscode.sh
@@ -21,17 +21,14 @@ main() {
     read -p "$prefix Enter [y|n]:" answer
     case $answer in
       [Yy])
-        if ! type -t brew
+        if [[ $(type -t brew) = "" ]] && [ ! -f /opt/homebrew/bin/brew ]
         then
-          if [[ ! -e /opt/homebrew/bin/brew ]]
-          then
             info "Cannot run this bootstrapper. Install required dependency first: Homebrew."
-          else
-            eval "$(/opt/homebrew/bin/brew shellenv)"
+        else
+          eval "$(/opt/homebrew/bin/brew shellenv)"
 
-            info "Installing VScode"; # -------------------------------------------------------------
-            brew install --cask visual-studio-code
-          fi
+          info "Installing VScode"; # -------------------------------------------------------------
+          brew install --cask visual-studio-code
         fi
         break;
       ;;
